local-AIDaemon-agentic-rag-adaptative/VERSIONE5/requirements.txt
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.24.3
